Prune needless .github/workflows (#6)
Closes #1

I remove source-web-tool-scripts.yal because the scope of this repo is
narrower than that of the webtool (which deals with interactive report).

I remove pr-check-file-encodings.yml because it's unconventional and
I'm not sure we need it. To stay lean I assume we don't need it and will
add it only if we later discover that we do.