Snapshot result datasets in full (#59)
* Results snapshots now show datasets in full

Closes #55

Each dataset under 40_Results is now captured in full by a
snpashot (in */_snaps/run_pacta.md).

All datasets are stored in a named list, and printed via a single
call to expect_snapshot(). I tested and confirm the result is the
same if instead we have multiple calls to expect_snapshot() -- one
per dataset. This makes sense as testthat creates one snapshot file
in */_snaps/file.md for each */test-file.R -- not one file per call
to expect_snapshot().

However we do it, the what I think we want is to be sure that the
test will fail if the results change in any row -- not just in the
first few rows that were printed before this commit. This is now
achieved by converting each tibble in the list to a vanilla
data.frame, as print.data.frame() prints all rows in a tibble.

* Snapshot the typeof columns

The type of some column may change and the data.frame
snapshot wouldn't detect it. This commit adds a snapshot
of the type of each column of each dataset.